uint -> uint256

+++ mocks/pumps/MockPump.sol
@@ -14,15 +14,11 @@ import "src/interfaces/pumps/IPump.sol";
 contract MockPump is IPump {
     bytes public lastData;
 
-    function update(uint[] calldata, bytes calldata data) external {
+    function update(uint256[] calldata, bytes calldata data) external {
         lastData = data;
     }
 
-    function read(address, bytes calldata)
-        external
-        view
-        returns (bytes memory data)
-    {
+    function read(address, bytes calldata) external view returns (bytes memory data) {
         return lastData;
     }
 }
