+++ script/deploy/Well.s.sol
@@ -23,7 +23,7 @@ contract DeployWell is Script, WellDeployer {
     IERC20 constant WETH = IERC20(0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2); // WETH9
 
     function run() external {
-        uint deployerPrivateKey = vm.envUint("PRIVATE_KEY");
+        uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
         vm.startBroadcast(deployerPrivateKey);
 
         // Tokens
